check for empty fields
