+++ templates/edit_info.html
@@ -6,7 +6,6 @@
         <title>{{title}}</title>
         <link rel="stylesheet" href="{{url_for('static', filename='css/styles.css')}}">
     </head>
-
     {% include 'header.html' %}
     
     <body>
@@ -16,6 +15,14 @@
                 <br/><br/>
                 <span class="instructions">Use the following form to edit a Student Record</span>
                 <br/><span class="error">{{error}}</span>
+                {% if error=='empty_fields' %}
+              <p class='error'>There are empty fields</p>
+                {% if empty_fields %}
+                    {% for empty_field in empty_fields %}
+                <p class='error'>{{empty_field.replace('_',' ').title()}} is empty!</p>
+                    {% endfor %}
+                {% endif %}
+                {% endif %}
 
                 <br/><br/>
                 <form 
@@ -31,11 +38,6 @@
 
         {% if page_type=="edit_student_details" %}
             <div class="container">
-                {% if error=="class_not_exist" %}
-                    <p class="error">
-                        New class provided does not exist!
-                    </p>
-                {% endif %}
                 <span class="highlight-blue">Update a Student Record</span>
                 <br/><br/>
                 <span class="instructions">Use the following form to update a Student Record</span>
@@ -51,12 +53,27 @@
                             <label for= "{{name}}">
                                 {{ name.replace("_", " ").title() }}:
                             </label>
-
-                            <input
+        
+                            {% if name=="student_class" %}
+                                <select name="student_class" id="student_class">
+                                    {% for class_ in class_list %}
+                                        {% if class_ == value %}
+                                            <option value="{{class_}}" selected>{{value}}</option>
+                                        {% else %}
+                                            <option value="{{class_}}">{{class_}}</option>
+                                        {% endif %}
+                                            
+                                    {% endfor %}
+                                </select>
+                            {% else %}
+                                <input
                                 id="{{name}}"
                                 type="text"
                                 name="{{name}}"
                                 value="{{value}}"/>
+                            {% endif %}
+                            
+                            
                             </td>
                         </tr>
                         {% endfor %}
@@ -117,8 +134,8 @@
 
             <p>Please click on the cca you want to edit</p>
             <ol>
-                {% for cca in record %}
-                <li><a href="/edit_membership?cca={{ cca.replace(' ', '+') }}&name={{ name.replace(' ', '+') }}">{{cca}}</li> <!-- so that when the page changes to page 3, we can carry over the cca name that the user wants to change, as well as their name--> 
+                {% for record in records %}
+                <li><a href="/edit_membership?cca={{ record['cca name'] }}&name={{ record['student name'] }}&role={{ record['role'] }}">{{ record["cca name"] }}</li> <!-- so that when the page changes to page 3, we can carry over the cca name that the user wants to change, as well as their name--> 
                 <!-- it works through the request.args. So the args would be a dictionary, something like {"cca": "comp", "name" = "xy"} -->
                 {% endfor %}
             </ol>
@@ -170,13 +187,16 @@
                             <label for="student_role">Student Role:</label>
                         </td>
         
-                        <td>
+                        <!-- <td>
                             <select name="student_role" id="student_role">
                               <option value="Member">Member</option>
                               <option value="Exco">Exco</option>
-                              <!-- <option value="mercedes">Mercedes</option>
-                              <option value="audi">Audi</option> -->
+                              <option value="President">President</option>
+                              <option value="Vice President">Vice President</option>
                             </select>
+                        </td> -->
+                        <td>
+                            <input id="student_role" type="text" name="student_role" value="{{ form_data['student_role'] }}">
                         </td>
                         </tr>
                         <td>
@@ -207,7 +227,7 @@
                 {% if error == "student_not_found" %}
                     <p class="error">Student does not exists! or Student has no activity records to edit</p>
                 {% endif %}
-                <span class="highlight-blue">Edit a Student activity Record</span>
+                <span class="highlight-blue">Edit a Student Activity Record</span>
                 <br/><br/>
                 <span class="instructions">Use the following form to edit a Student Record</span>
