+++ templates/index.html
@@ -16,13 +16,13 @@
         <li>Add Student's information</li>
         <center>
         <a href="/new_student">
-        <button>Add A New Student</button></a>
+        <button>Add A New Student Record</button></a>
 
         <a href="/new_student_cca">
-        <button>Add A New Student CCA</button></a>
+        <button>Add A New Student CCA Record</button></a>
         
         <a href="/new_student_activity">
-        <button>Add A New Student Activity</button></a>
+        <button>Add A New Student Activity Record</button></a>
         </center></li>
 
     </br></br>
@@ -44,13 +44,13 @@
         <button>View A Student's Profile</button></a>
 
         <a href="/view_by_class">
-        <button>View by class</button></a>
+        <button>View a class</button></a>
             
         <a href="/view_by_cca">
-        <button>View by CCA</button></a>
+        <button>View a CCA</button></a>
         
         <a href="/view_by_activity">
-        <button>View by Activity</button></a>
+        <button>View a Activity</button></a>
         </center>
 
     </br></br>
@@ -61,10 +61,10 @@
         <button>Edit Student's Details</button></a>
         
         <a href="/edit_membership">
-        <button>Edit Student's CCA</button></a>
+        <button>Edit Student's CCA Record</button></a>
         
         <a href="/edit_participation">
-        <button>Edit Student's Activity</button></a>
+        <button>Edit Student's Activity Record</button></a>
         </center>
 
     </br></br>
@@ -74,10 +74,10 @@
             <button>Delete Student Record</button>
 
             <a href='/delete_student_cca'>
-            <button>Delete Student CCA</button>
+            <button>Delete Student CCA Record</button>
 
             <a href='/delete_student_activity'>
-            <button>Delete Student Activity</button>
+            <button>Delete Student Activity Record</button>
 
             <a href='/delete_cca'>
             <button>Delete A CCA</button>
@@ -88,6 +88,4 @@
     </ol>
         </div>
     </body>
-
-    
 </html>
