+++ templates/new_cca_activity.html
@@ -72,8 +72,19 @@
 
     {% if page_type=='add_cca_fail' %}
     <div class='container'>
+        {% if error == 'already_exist' %}
         <p class="error">The following CCA already exists!</p>
-
+        {% endif %}
+        
+        {% if error == 'empty_fields' %}
+        <p class='error'>There are empty fields</p>
+        {% if empty_fields %}
+            {% for empty_field in empty_fields %}
+        <p class='error'>{{empty_field.replace('_',' ').title()}} is empty!</p>
+            {% endfor %}
+        {% endif %}
+        {% endif %}
+        
         {% for name, value in cca.items() %}
         {{ name.replace('_', " ").title()}}:{{value}}
     </br>
@@ -153,7 +164,21 @@
 
     {% if page_type=='add_activity_fail' %}
     <div class='container'>
+
+        {% if error == 'already_exist' %}
         <p class="error">The following Activity already exists!</p>
+        {% endif %}
+        
+        {% if error == 'empty_fields' %}
+        <p class='error'>There are empty fields</p>
+        {% if empty_fields %}
+            {% for empty_field in empty_fields %}
+        <p class='error'>{{empty_field.replace('_',' ').title()}} is empty!</p>
+            {% endfor %}
+        {% endif %}
+        {% endif %}
+
+        
 
         {% for name, value in activity.items() %}
         {{ name.replace('_', " ").title()}}:{{value}}
